MNT Add guard around SIGSTOP

+++ lute/tasks/task.py
@@ -159,7 +159,10 @@ class Task(ABC):
             self._report_to_executor(start_msg)
 
         # We stop process here so Executor can do any tasklet work if needed
-        os.kill(os.getpid(), signal.SIGSTOP)
+        if os.getenv("LUTE_CONFIGPATH") is not None:
+            # Guard w/ environment variable that is set only by Executor - don't
+            # SIGSTOP if Task is running without Executor
+            os.kill(os.getpid(), signal.SIGSTOP)
 
     def _signal_result(self) -> None:
         """Send the signal that results are ready along with the results."""
